Finish analysis
Adds a test for "bad actors" - those that were out of compliance in the selected month and also habitually have been out of compliance (they will get a pass under the memo but they shouldn't!)

Adds analysis of stack tests and maps facilities that "should've" stack-tested in the selected month (the bottom quartile for testing over the past 20 years).

Maps facilities that used the COVID code to indicate why they couldn't submit reports in the selected month.